requirements.txt
instagrapi
tqdm
python-dotenv
requests
Pillow
